+++ app/ExternalFiles/Extensions/Import_Export.php
@@ -35,11 +35,11 @@ class Import_Export extends Extension_Base {
 	protected string $name = 'import_export';
 
 	/**
-	 * The extension type.
+	 * The extension types.
 	 *
-	 * @var string
+	 * @var array<int,string>
 	 */
-	protected string $extension_type = 'import_dialog';
+	protected array $extension_types = array( 'import_dialog' );
 
 	/**
 	 * Instance of actual object.