+++ app/ExternalFiles/Extensions/Show_What_Will_Be_Done.php
@@ -30,11 +30,11 @@ class Show_What_Will_Be_Done extends Extension_Base {
 	protected string $name = 'show_what_will_be_done';
 
 	/**
-	 * The extension type.
+	 * The extension types.
 	 *
-	 * @var string
+	 * @var array<int,string>
 	 */
-	protected string $extension_type = 'import_dialog';
+	protected array $extension_types = array( 'import_dialog' );
 
 	/**
 	 * Instance of actual object.